pyproject.toml
[project]
name = "avito"
version = "4.0.1"
description = "Avito Backend trainee"
authors = [
    {name = "nikita",email = "nikitashelydko@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "databases (>=0.9.0,<0.10.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "ruff (>=0.14.6,<0.15.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
exclude = [
    ".git",
    ".venv",
    "migrations/versions",
    "__pycache__",
    ".idea/"
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"